Added a prototype 7th order RK integrator
